Delete writingdex controller
