+++ app.py
@@ -1,5 +1,4 @@
 from flask_app.controllers import (
-    writingdex_controller,
     index_controller,
     writing_controller, 
     creature_controller,
