+++ flask_app/controllers/writing_controller.py
@@ -1,6 +1,12 @@
 from flask_app import app
-from flask import render_template, redirect
+from flask import render_template, redirect, request
 from flask_app.models import writing
+
+@app.route('/w')
+@app.route('/w/')
+def writings():
+    writings = writing.Writing.get_all()
+    return render_template("writingdex.html", writings=writings)
 
 @app.route('/w/<writing_url>')
 def view_writing( writing_url ):
@@ -14,7 +20,6 @@ def view_writing( writing_url ):
         if next_char == "-":
             writing_time_s = writing_url[:i]
             break
-    
 
     # safety checks for handling our url string
     # just in case someone manually types in a url
@@ -32,3 +37,19 @@ def view_writing( writing_url ):
         return redirect('/w')
 
     return render_template("view-writing.html", writing=one_writing )
+
+@app.route('/w/create-writing')
+def create_writing():
+    return render_template("create-writing.html")
+
+@app.route('/w/save-writing', methods=["POST"])
+def save_writing():
+    data = {
+        "name": request.form["name"],
+        "writing": request.form["writing"],
+        "description": request.form["description"],
+        "time_s": request.form["time_s"],
+    }
+
+    writing.Writing.save( data )
+    return redirect('/w/create-writing')
